refactor: move ProductExportOptions to FileStorage module
Relocates ProductExportOptions from Domain/Features/Configuration to
Application/Features/FileStorage where its only consumers live. Aligns
with the project's Vertical Slice convention — every other single-file
options class (Article, Leaflet, OrgChart, MeetingTasks, KnowledgeBase)
sits inside its own feature slice.

Pure refactor: class body, defaults, section key ("ProductExportOptions"),
and DI binding semantics are byte-for-byte unchanged. Only namespace
and using directives moved.
