+++ backend/test/Anela.Heblo.Tests/Features/FileStorage/FileStorageModuleTests.cs
@@ -1,6 +1,5 @@
 using Anela.Heblo.Application.Features.FileStorage;
 using Anela.Heblo.Application.Features.FileStorage.Infrastructure;
-using Anela.Heblo.Domain.Features.Configuration;
 using Anela.Heblo.Domain.Features.FileStorage;
 using Anela.Heblo.Xcc.Telemetry;
 using Microsoft.Extensions.Configuration;
