+++ backend/test/Anela.Heblo.Tests/Features/FileStorage/Infrastructure/DownloadResilienceServiceTests.cs
@@ -1,5 +1,5 @@
 using Anela.Heblo.Application.Features.FileStorage.Infrastructure;
-using Anela.Heblo.Domain.Features.Configuration;
+using Anela.Heblo.Application.Features.FileStorage;
 using Anela.Heblo.Xcc.Telemetry;
 using FluentAssertions;
 using Microsoft.Extensions.DependencyInjection;
